Move admin pages to use appropriate folders
