+++ pages/admin/index.js
@@ -1,14 +1,14 @@
 import React from "react";
-import Layout from "../src/components/Layout";
-import propsWithContainer from "../src/middleware/propsWithContainer";
-import verifyAdminToken from "../src/usecases/verifyAdminToken";
-import { GridRow, GridColumn } from "../src/components/Grid";
-import Heading from "../src/components/Heading";
-import ActionLink from "../src/components/ActionLink";
-import TrustsTable from "../src/components/TrustsTable";
-import Text from "../src/components/Text";
+import Layout from "../../src/components/Layout";
+import propsWithContainer from "../../src/middleware/propsWithContainer";
+import verifyAdminToken from "../../src/usecases/verifyAdminToken";
+import { GridRow, GridColumn } from "../../src/components/Grid";
+import Heading from "../../src/components/Heading";
+import ActionLink from "../../src/components/ActionLink";
+import TrustsTable from "../../src/components/TrustsTable";
+import Text from "../../src/components/Text";
 import Error from "next/error";
-import { ADMIN } from "../src/helpers/userTypes";
+import { ADMIN } from "../../src/helpers/userTypes";
 
 const Admin = ({ trusts, error }) => {
   if (error) {
